requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
idna==3.4
isodate==0.6.1
lxml==4.9.1
platformdirs==2.5.4
pytz==2022.6
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
